+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add new `Array#findByPath(obj)` signature
 * Add `Object.sizeof(input)` to get the size of a variable in bytes
 * Add `Cache` class
+* Allow passing a descriptor to `Blast.defineValue`
 
 ## 0.5.11 (2018-07-14)